refactor: Fix bug

+++ frontend/lib/pages/tab_3/manage_racket/detail_racket/dep_2_racket_history/UserRacketHistory.model.dart
@@ -100,7 +100,7 @@ class ListElement {
   String replacementGripType;
   int overgripNum;
   String racketBalanceType;
-  int racketBalanceVal;
+  double racketBalanceVal;
   double mainGutLbTension;
   double crossGutLbTension;
   String gutCompanyName;
